For do_py(), moved the definition of the run() command earlier.
This is so that it will allow the user to run a Python script directly from the Cmd prompt even before an interactive Python shell has ever been invoked via:
    py run('script.py')

This fixes a bug where it wouldn't allow that to occur until the user entered just "py" at the Cmd prompt.

Also added a trivial Python script script.py to the example directory for the purposes of testing script execution.