Fixed pipe to shell bug on Python 3.x.
Fixed implementation of how command output can get piped to OS shell commands so that it works in Python 3 as well as Python 2.  Also added a unit test to verify that this functionality is behaving as intended.

All I can say is that the subprocess module was horrible for Python 3 prior to 3.5.  Getting it to play nicely with unicode strings for stdin/stdout isn't easy.