Made argument parsing more backward compatible by default.
Added a global variable BACKWARD_COMPATIBLE_PARSING which defaults to True.  If this is set to True, then if shelx.split() throws an exception when splitting up the arguments, it falls back to splitting them up using str.split() instead.

If this is False, it "throws" an exception and prints an error message at the console with details on the error.

TODO: Figure out a way to make this settable from within the Cmd class, but accessible within the options decorator.  I prefer it False for a stricter grammar to help prevent user error.  But I imagine there are people who prefer it True for backward compatibility.  It exists for now in this "non-settable" hard-coded state because I did not intend to break backward compatibility when fixing the bug related to quoted arguments for which shelx.split() was brought in to deal with.