Experiment with different modes of parsing command-line arguments for option commands.
This is an experiment with usin non-POSIX parsing of the command line with shlex before passing it to the optparse OptionParser.  The posix vs non-posix setitng primarily effects how quotes and escape characters are dealt with.

I'm experimenting with various options to see what may feel the most natural for the majority of end users.