Moved preloop() and postloop() calls from _cmdloop() to cmdloop().
In order to fix a problem with do_load() trying to do more than it should, refactored to move the preloop() and postloop() calls so they are not called when loading a script.

This is so a script is simply a way to automate running several commands in a row and makes it safe for preloop() to create threads, connect to servers, and do other stateful things.