Added optional parameter to cmd2.Cmd.__init() to enable an "ipy" command.
If IPython is installed AND the __init__ initilizer is passed a "use_ipython" keyword argument set to "True", then and "ipy" command will be present.

This ipy command starts an interactive IPython shell embedded within the cmd2 application.  This IPython shell has access to all of the local variables (e.g. self), but any modifications made within it will NOT persist once it is exited.  This IPython shell doesn't have tight integration with the Python interpreter running the cmd2 application like the Python interpreter available with the "py" command.  However, due to features such as tab-completion and built-in help, the IPython shell is a lot more efficient for developers to use to debug their cmd2 applications - the introspection on self is extremely useful for getting to the bottom of thorny debugging problems.