Removed inclusion and use of doctest since all doctests converted to pytest.
Since all doctests were converted to pytest unit tests in a previous commmit, this commit just removes the inclusion and usage of doctest.  Now if cmd2.py is run as a main application, it just starts a bare-bones cmd2.Cmd() application with only built-in functionality instead of running the doctests which no longer exist.