Abbreviations are no longer accepted for multiline commands
Due to the way the parsing logic works for multiline commands, abbreviations didn't function properly with mutliline commands.  So to avoid confusion, this commit deals with this issue by simply disallowing abbreviations for multiline commands altogether.

A warning has been added to the section on abbreviations in the documentation to hopefully make this clear for users.