Added "exclude_from_help" list as a cmd2.Cmd public attribute
Added a list of commands to explicitly exclude displaying in the help menu of commands which detailed help can be retrieved for.

By default, only "do_eof" is in this list since that isn't a command which is ever intended for an end user to directly enter on the command line.

Also updated unit tests to reflect new default help menu.