Removed redundant list command
The extra functionality of listing a span (slice) of history items present in the list command has been incorporated into the existing history command and the list command has been removed.

Unit tests have been updated accordingly.

In the process a bug was fixed that was causing a crash if and end was supplied to the span but not a start.