Updated cmdenvironment command
It no longer displays settable parameters since this functionality was redundant since those are displayed with either set or show.

It now shows additional global settings for parsing @options commands.

In the future a few more things should probably be displayed here related to parameters which are not settable at runtime.