Changed default value for USE_ARG_LIST global to True
Now by default all @options commands get passed a list of argument strings instead of a single argument string.

This is a much easier and more robust behavior to deal with.  Additionally, command-line arguments are intelligently separated based on location of quotes to group things into a single argument.

WARNING:  This change breaks backward compatibility for older applicaitons based on cmd2.  To change the behavior to the way it used to be, add the following code to the __init__() method of our class derived from cmd2.Cmd:
    cmd2.set_use_arg_list(False)

This change really does make it easier for developers new to using cmd2 however.  It is to the point where I create all custom commands with @options, even if I use an empty list for the options because the argument parsing is just much better this way.