Removed some help_* methods
I originally created separate help_* methods for various built-in do_* methods so that:
- The Sphinx docs could use the do_* docstring which could have extra reStructuredText markup embedded
- The built-in help command could use the text from the help_* method

This seemed like a good idea at the time, but I changed my mind - I was actually adding a maintenance headache.
- And the payoff was small

So I have now switched to using the do_* docstring for both the built-in help and Sphinx documentation.
- With some very minor usage of reStructuredText markup that doesn't look bad when displaying built-in help