First attempt at improving how command output gets piped to a shell command
Now a real pipe is created to a subprocess.  This has many advantages and should "just work" like intended with all commands.

One downside is to work properly on Python 2.7, it requires the subprocess32 module which is the subprocess module from Python 3.2 backported to Python 2.7.  Another downside, is that unit testing the feature is now more difficult.

This still needs to be tested for compatibility across all OSes and supported versions of Python.

The user needs to be careful if designing multi-threaded cmd2 applications that do command processing in other threads and those threads can make calls to self.stdout.write to put in a try/except to catch Broken Pipe errors which can occur for long running commands if the user closes the shell subprocess before the command is finished.