Replaced direct calls to self.stdout.write() with calls to self.poutput().
Modified implementation of poutput() to accept an optional "end" argument which can change the ending to something other than a newline or simply suppress adding a newline.  Also added a try/except to catch BrokenPipeError exceptions which can occupoutputr if the subprocess output is being piped to closes before the command piping output to it is finished.

Updated install docs to note that when using Python 2.7, the subprocess32 module should also be installed.