Provide method to run multiple commands w/o a cmdloop.
runcmds_plus_hooks can accept multiple commands process the command queue to
 deal with subsequent commands loaded from scripts without requiring a command
 loop. This better supports a one-off batch processing scenario.

Also fixed the insertion order of commands placed in the command queue by load
and _relative_load so that script commands are run in the expected order.

Minor tweak to setup instructions in CONTRIBUTING.md to include pyperclip in
prerequisites.