help command temporarily redirects sys.stdout and sys.stderr to self.stdout for argparse commands
In order to make "help" behave more consistently for decorated and undecorated commands, argparse output is temporarily redirected to self.stdout.  So doing "help history" is similar to "help load".

However, when using the "-h" with argparse commands without using the "help" command, the output from argparse isn't redirected to self.stdout.  Fixing this would be rather difficult and would essentially involve creating a pyparsing rule to detect it at the parser level.