First past at unit test for persistent history feature
Added pexpect to modules required for running unit tests.

This opens the door for carefully crafted complex unit tests to verify intricate behavior.  Tests like this are somewhat painful to write and slow to execute.  However, they can enable testing complicated interactive behavior that we otherwise probably would not be able to test.