Converted a few class variables into instance variables
Now that users can nest instances of cmd.Cmd2 to support creating sub-menus, we should need to be more careful about class vs instance variables to prevent potential problems.

This converts the following former class variables into instance variables:
- multiline_commands
- shortcuts
- exclude_from_help
- exclude_from_history

In the process, a couple camelCase variable names got converted to pep8_compliant names.

There may be a few other class variables which should be converted to instance variables.  But at the very least, this is a good start.

This closes #273.