Skip test_default_to_shell_completion unit test on Windows
Python's built-in cmd module (and cmd2 which inherits from it) parses the command line and has relatively strict rules when parsing the beginning of the line looking for a "command".  This command can must consist of ASCII letters and digits.  The parsing logic applies prior to the default_to_shell flag attempting to run this "command" as a shell command.

A problem occurs on Windows in that nearly all executables have a "." in their name.  cmd/cmd2 will view everything before the "." as being the command.

On Windows, I would strongly encourage the use of "!" or "shell" to run shell commands and would recommend against using the "default_to_shell" flag unless you are working within the Windows Subsystem for Linux (WSL).