Make alias checking and command parsing use the same regex
Provide a new is_valid_command() method on StatementParser to determine whether a string of characters could be a valid command. That means it can’t include any redirection, quote chars, whitespace, or terminator characters. This method is used when someone tries to create an alias, to ensure when we try and parse the alias that it will actually parse.

This nicely encapsulates and standardizes all the logic for parsing and expansion into the StatementParser class.

Also fix a bug in the regex to match valid command names, and add a bunch of new unit tests to ensure the bug stays fixed.