Fixed a couple potential crashes on opening files
Fixed crashes that occur when attempting to open a file in a non-existent directory or a when the filename is too long.

Specifically fixed this when redirecting output to a file and when saving a transcript based on the history.

Also added a couple unit tests related to the fixes.