Added back bash_complete() helper function to argcomplete_bridge.py
This is intended to be used once a new example is added to show how to use the code present in argcomplete_bridge.

This was previously used by the subcommands.py example, but I deleted the argcomplete_bridge code from there to simplify that example and make it focused on subcommands.