Fix flake8 issues
This commit contains a very large number of trivial changes in order to fix flake8 errors and warnings.  Predominantly these are whitespace changes.

Additionally, the build for Python 3.7 on TravisCI has been tweaked to fail if there are any flake8 errors using the following commandline:
* flake8 . --count --ignore=E252 --max-complexity=31 --max-line-length=127 --show-source --statistics
NOTE: In the future the max cyclomatic complexity should be lowered, but some improvements need to be made first.

One flake8 error is being ignored entirely:
* E252 missing whitespace around parameter equals
    * ignored because it doesn't correctly deal with default argument values after a type hint

A few flake8 errors are being selectively ignored in certain files:
* C901 fuction is too complex
    * ignored in argparse_completer.py because the complex code is an override of argparse complexity
* E302 expected 2 blank lines after ...
    * ignored in all unit test files for convenience
* F401 module imported but unused
    * ignored in cmd2/__init__.py because imports are for convenience of cmd2 developers and backwards compatibility
* F821 undefined name
    * ignored in cmd2 script files which are intended to run only within cmd2 applications via pyscript where "app" and "cmd" are defined