Fix a bug discovered during manual testing
I found that at least with certain versions of Python and OSes, if I had a previous text-based readline history, an unhandled UnpicklingError exception could occur.  So now we catch that and several other possible errors which can theoretically occur during unpickling and just move on with an empty history.  The old file will get overwritten with one in the new format when the application exits.