update postparsing hooks
Previously redirection information was lost by the postparsing hooks
add_whitespace_hook() and downcase_hook(). This was fine for this simple
shell, but sets a bad example for more complicated use cases.

To fix this, when parsing the new Statement object, include the
`post_command` attribute from the original Statement.