Add 'history_includes_scripts' settable
At least in most of my use cases, the individual commands of a script
executed by the command line should not end up in the history.  If I go
back in history, I want to find the commands I typed (like the
run_script), and not the commands contained in the script I executed,
or even the scripts that this script might execute itself.

This patch introduces a new settable 'history_includes_scripts'.  It is
kept as 'true' by default to preserve the legacy behavior of existing
cmd2 versions.  However, it can be set to 'false' to make run_script
not add to the history by default.  Individual run_script executions
can still use the --history option to make them override the default
and add to the history.

Signed-off-by: Harald Welte <laforge@osmocom.org>