Undo backwards-incompatible change in naming of cmd2.ansi.Cursor method. (#1430)
It just wasn't worth making a backwards-incompatible change which might impact people, especially since we are likely to replace the ansi module with one based on rich in the near future.

Also updated the Makefile to also install prettier using npm and added the package.json file to Git to support this.