Fix Codecov code coverage analysis from test data (#1434)
This PR does a few things in pursuit of fixing our integrated code coverage analysis provided by Codecov:

It removes the last usage of invoke from our GitHub Actions CI scripts. Now we rely purely on uv for everything.
Uploading of data to Codecov is now done using a private token which is stored in GitHub secrets for this repo.
Both test data and code coverage are now uploaded.