Ruggedize cmd2's runtime type annotation validation (#1442)
* Change cmd2's runtime type annotation validation to be based on typing.get_type_hints

Previously it was based on inspect.signature. The problem is that to Python 3.10, the inspect module doesn't have a safe way of evaluating type annotations that works equivalently both in the presence or absence of "from __future__ import annotations". Hence, any attempt at using that in an app would break cmd2.

This change adds a get_types() helper function to the cmd2.utils module which uses typing.get_type_hints() to do the introspection in a safer way.

* Fix for Python 3.9 since types.NoneType doesn't exist until 3.10

* Added unit tests for new utils.get_types function

* Add a note to the CHANGELOG

* Added a new test file to cover the case representing the whole reason for this change

I first verified that this test fails on the current master branch

* Switched CHANGELOG type to Bug Fix