Add mypy type checking for Python 3.14 and use open() instead of codecs.open() due to deprecation in 3.14
The first release candidate for Python 3.14 was released on 2025-07-22 and I did some local testing.

The first thing I realized is that mypy now works well with 3.14 and so we should add type checking on that platform to our GitLab Actions.

The second thing I noticed is that when we ran tests with pytests, we were getting a bunch of warnings due to `codecs.open()` being deprecated.  Since at least Python 3.6 there has been no need for that as `open()` does everything it can do plus more.