Simplify subcommand example within argparse_example.py (#1498)
Simplify subcommand example within argparse_example.py to be a simple calculate command with add and subtract subcommands.

The original subcommand example was overly complicated. This replaces it with a much simpler one that is easier to digest and understand.

Also switched to using the cmd2.as_subcommand_to decorator.